requirements.txt
amqp==5.0.6
aniso8601==7.0.0
asgiref==3.4.1
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-celery-beat==2.2.1
django-ckeditor==6.1.0
django-filter==21.1
django-graphql-jwt==0.3.4
django-js-asset==1.2.2
django-redis==5.0.0
django-taggit==1.5.1
django-timezone-field==4.2.1
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
kombu==5.1.0
Pillow==8.3.2
promise==2.3
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
redis==3.5.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
